recipe.ts adjusted

+++ src/datasources/recipe.ts
@@ -68,18 +68,20 @@ class Recipe extends DataSource {
     recipeId: string
   ): Promise<RecipeModel | null> {
     if (!this.context.user) return null;
-    const recipe = await dbRecipe.findOne({
-      where: { id: recipeId, ownerId: this.context.user.id },
-    });
+    const recipe = await dbRecipe.findOne({ where: { id: recipeId } });
     if (!recipe) return null;
+    if (recipe.ownerId !== this.context.user.id) return null;
     return await recipe.update(updatedRecipe);
   }
 
   async deleteRecipe(recipeId: string): Promise<boolean> {
     if (!this.context.user) return false;
+    const recipe = await dbRecipe.findOne({ where: { id: recipeId } });
+    if (!recipe) return false;
+    if (recipe.ownerId !== this.context.user.id) return false;
     return (
       (await dbRecipe.destroy({
-        where: { id: recipeId, ownerId: this.context.user.id },
+        where: { id: recipeId },
       })) === 1
     );
   }
